sift-v1/requirements.txt
Flask==2.0.1
cryptography==3.4.7
pycryptodome==3.10.1
pytest==6.2.4
requests==2.26.0